SVG: Increase hit-testing precision on strokes
Skia takes a scale-factor when generating complex strokes (dashed, etc.)
as an optimization.  When rendering, we passed a scale-factor derived
from the CTM, but when hit-testing we were passing a hard-coded value.
As a result, the stroke generated for hit-testing was noticeably
different between hit-testing and rendering.

This change passes the CTM through to the hit-testing code, so that the
same scale factor will be used for hit-testing as it was for rendering.

Bug: 964614
Change-Id: I8dfcb3ade86b9e7b63c7415baab5814ac2286ce8